# swagger.json
{"openapi":"3.0.0","info":{"version":"1.0.0","title":"My subs","description":"API in NodeJs using TypeScript"},"servers":[{"url":"https://my-subs-api.herokuapp.com/","description":"Production"},{"url":"http://localhost:4000/","description":"Local"}],"components":{"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}},"schemas":{"IUser":{"type":"object","required":["_id","email","password","rol","tokens"],"properties":{"_id":{"type":"integer"},"email":{"type":"string"},"name":{"type":"string"},"lastname":{"type":"string"},"password":{"type":"string"},"tokens":{"type":"array","items":{"type":"string"}},"deletedAt":{"type":"string"}}},"IErrorResponse":{"type":"object","required":["status","code","message"],"properties":{"status":{"type":"string"},"code":{"type":"string"},"message":{"type":"string"}}},"ISubscription":{"type":"object","required":["_id","name","cost","paidPeriod","isPaid"],"properties":{"_id":{"type":"integer"},"name":{"type":"string"},"paidPeriod":{"type":"string","enum":["WEEK","MONTH","YEAR"]},"isPaid":{"type":"boolean"},"note":{"type":"string"},"deletedAt":{"type":"date"}}}}},"security":[{"bearerAuth":[]}],"paths":{"/user":{"get":{"description":"Get all users","tags":["User"],"responses":{"200":{"description":"List of users","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/IUser"}}}}},"400":{"description":"ERROR - Bad Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/IErrorResponse"}}}},"401":{"description":"ERROR - Error not authenticated","content":{"application/json":{"schema":{"$ref":"#/components/schemas/IErrorResponse"}}}},"500":{"description":"ERROR - Server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/IErrorResponse"}}}}}},"post":{"description":"Register user","tags":["User"],"requestBody":{"content":{"application/json":{"schema":{"properties":{"email":{"type":"string"},"name":{"type":"string"},"lastname":{"type":"string"},"password":{"type":"string"}}}}}},"responses":{"200":{"description":"The user has been registered","content":{"application/json":{"schema":{"type":"object","properties":{"authToken":{"type":"string"}}}}}},"400":{"description":"ERROR - Bad Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/IErrorResponse"}}}},"500":{"description":"ERROR - Server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/IErrorResponse"}}}}}}},"/user/{id}":{"delete":{"description":"Logic delete user","tags":["User"],"parameters":[{"name":"id","in":"path","description":"ID of user to delete","required":true,"schema":{"type":"string"},"style":"simple"}],"responses":{"200":{"description":"The user has been deleted","content":{"application/json":{"schema":{"type":"object","properties":{"message":{"type":"string"}}}}}},"400":{"description":"ERROR - Bad Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/IErrorResponse"}}}},"401":{"description":"ERROR - Error not authenticated","content":{"application/json":{"schema":{"$ref":"#/components/schemas/IErrorResponse"}}}},"403":{"description":"ERROR - Error not enouth permisions","content":{"application/json":{"schema":{"$ref":"#/components/schemas/IErrorResponse"}}}},"500":{"description":"ERROR - Server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/IErrorResponse"}}}}}},"patch":{"description":"Updates one user","tags":["User"],"parameters":[{"name":"id","in":"path","description":"ID of user to update","required":true,"schema":{"type":"string"},"style":"simple"}],"responses":{"200":{"description":"The user has been updated","content":{"application/json":{"schema":{"type":"object","properties":{"message":{"type":"string"}}}}}},"400":{"description":"ERROR - Bad Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/IErrorResponse"}}}},"401":{"description":"ERROR - Error not authenticated","content":{"application/json":{"schema":{"$ref":"#/components/schemas/IErrorResponse"}}}},"403":{"description":"ERROR - Error not enouth permisions","content":{"application/json":{"schema":{"$ref":"#/components/schemas/IErrorResponse"}}}},"500":{"description":"ERROR - Server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/IErrorResponse"}}}}}}},"/user/login":{"post":{"description":"Login user","tags":["User"],"requestBody":{"content":{"application/json":{"schema":{"properties":{"email":{"type":"string"},"password":{"type":"string"}}}}}},"responses":{"200":{"description":"The user has been logged","content":{"application/json":{"schema":{"type":"object","properties":{"authToken":{"type":"string"}}}}}},"400":{"description":"ERROR - Bad Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/IErrorResponse"}}}},"500":{"description":"ERROR - Server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/IErrorResponse"}}}}}}},"/user/logout":{"post":{"description":"Logout user","tags":["User"],"responses":{"200":{"description":"Logout successfully","content":{"application/json":{"schema":{"type":"object","properties":{"message":{"type":"string"}}}}}},"400":{"description":"ERROR - Bad Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/IErrorResponse"}}}},"500":{"description":"ERROR - Server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/IErrorResponse"}}}}}}},"/user/auth":{"get":{"description":"Get user auth","tags":["User"],"responses":{"200":{"description":"User auth successfully","content":{"application/json":{"schema":{"type":"object","properties":{"$ref":"#/components/schemas/IUser"}}}}},"400":{"description":"ERROR - Bad Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/IErrorResponse"}}}},"401":{"description":"ERROR - Error not authenticated","content":{"application/json":{"schema":{"$ref":"#/components/schemas/IErrorResponse"}}}},"500":{"description":"ERROR - Server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/IErrorResponse"}}}}}}}}}